+++ components/landing/SubscribeBanner.tsx
@@ -1,0 +1,52 @@
+import Link from "next/link";
+import React from "react";
+import {
+  AnnotationIcon,
+  GlobeAltIcon,
+  LightningBoltIcon,
+  ScaleIcon,
+} from "@heroicons/react/outline";
+import ContactForms from "../ContactForms";
+import Testimonial, { TestimonialProps } from "../landing/Testimonial";
+import { Button } from "../ui/Button";
+
+const SubscribeBanner = () => {
+  return (
+    <div>
+      <div className="grid grid-cols-1 sm:grid-cols-2 bg-secondary-third my-4 p-20">
+        <div className="p-12">
+          <p className="text-primary font-medium text-4xl">
+            Don't believe your college's marketing? In one project-based class
+            you can learn the fundamentals of coding.
+          </p>
+        </div>
+
+        <div className="items-center grid sm:grid-cols-2 grid-cols-1 gap-4 p-4">
+          <div className="bg-white flex sm:flex-row items-center rounded-lg sm:cols-span-2 sm: col-span-1">
+            <input
+              id="bootcamper"
+              type="text"
+              autoComplete="off"
+              className={`text-left p-2 border rounded-md shadow-md focus:outline-none focus:ring-indigo-500 text-md lg:text-md text-primary placeholder-yellow-700 w-full sm:w-96`}
+              placeholder="Email"
+            />
+          </div>
+          <div>
+            <div className="sm:justify-center lg:justify-start">
+              <div className="rounded-md shadow">
+                <a
+                  href="/courses"
+                  className="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-white bg-primary hover:bg-yellow-700 md:py-2 md:text-lg md:px-10"
+                >
+                  Enroll Now
+                </a>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default SubscribeBanner;
