+++ pages/index.tsx
@@ -104,7 +104,7 @@ const HomePage = () => {
 
             <main className="mt-10 mx-auto max-w-7xl px-4 sm:mt-12 sm:px-6 md:mt-16 lg:mt-20 lg:px-8 xl:mt-28">
               <div className="sm:text-center lg:text-left">
-                <h1 className="text-4xl tracking-tight font-extrabold text-gray-900 sm:text-5xl md:text-6xl text-center">
+                <h1 className="text-4xl tracking-tight font-extrabold text-gray-900 sm:text-5xl md:text-6xl">
                   <span className="">Learn new </span>{" "}
                   <span className=" text-yellow-600">skills</span>{" "}
                   <span className="">to </span>{" "}
